fixing columns mappings

+++ src/TitanicApi.Model/Model/Entities/PassengerEntity.cs
@@ -13,11 +13,11 @@ public class PassengerEntity : ITableEntity
     public int Pclass { get; set; }
     public string Name { get; set; }
     public string Sex { get; set; }
-    public decimal? Age { get; set; }
+    public int? Age { get; set; }
     public int SibSp { get; set; }
     public int Parch { get; set; }
     public string Ticket { get; set; }
-    public decimal Fare { get; set; }
+    public double Fare { get; set; }
     public string Cabin { get; set; }
     public string Embarked { get; set; }
 }
